flask-api/test_mysql_connection.py
import os
import mysql.connector
from dotenv import load_dotenv

load_dotenv()

try:
    conn = mysql.connector.connect(
        host=os.getenv("MYSQL_HOST"),
        user=os.getenv("MYSQL_USER"),
        password=os.getenv("MYSQL_PASSWORD"),
        database=os.getenv("MYSQL_DB"),
        port=int(os.getenv("MYSQL_PORT", 3306)),
    )
    print("✅ Kết nối MySQL thành công!")
    conn.close()
except Exception as e:  
    print("❌ Kết nối MySQL thất bại:", e)
