flask-api/requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
mysql-connector-python
marshmallow
marshmallow-sqlalchemy
Werkzeug
